refac: abstract company portfolios comp
