+++ libs/features/src/lib/project/components/organisms/company-portfolios/company-portfolios.spec.tsx
@@ -1,0 +1,12 @@
+import { render } from '@testing-library/react';
+
+import CompanyPortfolios from './company-portfolios';
+
+describe('CompanyPortfolios', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(
+      <CompanyPortfolios />
+    );
+    expect(baseElement).toBeTruthy();
+  });
+});
