+++ libs/features/src/lib/project/components/organisms/index.ts
@@ -1,0 +1,2 @@
+export * from './project-card/project-card';
+export * from './company-portfolios/company-portfolios';
