--- libs/features/src/lib/project/components/organisms/index.tsx
@@ -1,1 +1,0 @@
-export * from "./project-card/project-card";
